+++ foodway/app/src/main/java/com/example/foodway/model/Category.kt
@@ -1,0 +1,6 @@
+package com.example.foodway.model
+
+data class Category (
+    val urlImage: String,
+    val categoryName: String
+)
