+++ foodway/app/src/main/java/com/example/foodway/model/SignUpEstablishment.kt
@@ -1,0 +1,12 @@
+package com.example.foodway.model
+
+data class SignUpEstablishment(
+    val fantasyName: String,
+    val responsible: String,
+    val email: String,
+    val password: String,
+    val cnpj: Int,
+    val cep: Int,
+    val number: String,
+    val categories: List<Category>
+)
