+++ foodway/app/src/main/java/com/example/foodway/view/profileCustomer/ProfileCustomerActivity.kt
@@ -1,0 +1,59 @@
+package com.example.foodway.view.profileCustomer
+
+import android.os.Bundle
+import androidx.activity.ComponentActivity
+import androidx.activity.compose.setContent
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Surface
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.tooling.preview.Preview
+import com.example.foodway.ui.theme.FoodwayTheme
+import com.example.foodway.view.components.NavBarComponent
+
+class ProfileCustomer : ComponentActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContent {
+            ProfileCustomerScreen()
+        }
+    }
+}
+
+@Composable
+fun ProfileCustomerScreen() {
+    FoodwayTheme {
+        Surface(
+            modifier = Modifier.fillMaxSize(),
+            color = MaterialTheme.colorScheme.background
+        ) {
+            Column(
+                verticalArrangement = Arrangement.SpaceBetween
+            ) {
+                WelcomeProfile()
+                Row(
+                    modifier = Modifier.fillMaxWidth(),
+                    horizontalArrangement = Arrangement.Center
+                ) {
+                    CardInfoUser()
+                }
+                Column {
+                    RecentCard()
+                    FavoriteCard()
+                }
+                NavBarComponent()
+            }
+        }
+    }
+}
+
+@Preview(showBackground = true)
+@Composable
+fun DefaultPreview() {
+    ProfileCustomerScreen()
+}
