+++ foodway/app/src/main/java/com/example/foodway/view/signUp/establishment/Input.kt
@@ -1,0 +1,19 @@
+package com.example.foodway.view.signUp.establishment
+
+import androidx.compose.foundation.layout.Box
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.tooling.preview.Preview
+
+@Composable
+fun Input() {
+    Box {
+
+    }
+}
+
+
+@Preview(showBackground = true)
+@Composable
+fun InputPreview() {
+    Input()
+}
