+++ foodway/app/src/main/java/com/example/foodway/view/signUp/establishment/ScreenBorder.kt
@@ -1,0 +1,18 @@
+package com.example.foodway.view.signUp.establishment
+
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.tooling.preview.Preview
+
+@Composable
+fun ScreenBorder() {
+
+}
+
+
+@Composable
+@Preview(showBackground = true)
+fun ScreenBorderPreview() {
+    ScreenBorder()
+}
+
+
